+++ hutool-core/src/test/java/cn/hutool/core/codec/Base64Test.java
@@ -18,7 +18,7 @@ public class Base64Test {
 		String a = "伦家是一个非常长的字符串66";
 		String encode = Base64.encode(a);
 		Assert.assertEquals("5Lym5a625piv5LiA5Liq6Z2e5bi46ZW/55qE5a2X56ym5LiyNjY=", encode);
-
+		
 		String decodeStr = Base64.decodeStr(encode);
 		Assert.assertEquals(a, decodeStr);
 	}